feat(goap): implement missing plan tasks (S1.2, W2.3, W2.6, #837)
Land high-value items from GOAP_CODEBASE_IMPROVEMENTS via swarm:

- #837: restore public fuzzy_match rustdoc + doctests
- S1.2/ADR-074: CacheKey includes TaskContext language/framework/
  complexity/normalized tags; wire retrieve_relevant_context
- W2.3: build-rust accepts do-memory-* hyphenated packages
- W2.6: split production sources over 500 LOC
- S1.1a/D3.2: fail-closed execute_agent_code docs; fix README

Update plans/, CHANGELOG, ADR-074 partial status.
