+++ memory-core/src/memory/checkpoint/operations.rs
@@ -364,142 +364,9 @@ async fn get_relevant_heuristics(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::types::{TaskContext, TaskType};
-
-    #[tokio::test]
-    async fn test_checkpoint_episode_not_found() {
-        let memory = SelfLearningMemory::new();
-        let result = checkpoint_episode(&memory, Uuid::new_v4(), "test".to_string()).await;
-        assert!(result.is_err());
-    }
-
-    #[tokio::test]
-    async fn test_get_handoff_pack_not_found() {
-        let memory = SelfLearningMemory::new();
-        let result = get_handoff_pack(&memory, Uuid::new_v4()).await;
-        assert!(result.is_err());
-    }
-
-    #[tokio::test]
-    async fn test_checkpoint_completed_episode() {
-        use crate::episode::ExecutionStep;
-        use crate::memory::MemoryConfig;
-        use crate::types::ExecutionResult;
-
-        let test_config = MemoryConfig {
-            quality_threshold: 0.3,
-            ..Default::default()
-        };
-        let memory = SelfLearningMemory::with_config(test_config);
-
-        let episode_id = memory
-            .start_episode(
-                "Test task".to_string(),
-                TaskContext::default(),
-                TaskType::Testing,
-            )
-            .await;
-
-        let mut step = ExecutionStep::new(1, "test_tool".to_string(), "test action".to_string());
-        step.result = Some(ExecutionResult::Success {
-            output: "test output".to_string(),
-        });
-        memory.log_step(episode_id, step).await;
-
-        memory
-            .complete_episode(
-                episode_id,
-                crate::types::TaskOutcome::Success {
-                    verdict: "Done".to_string(),
-                    artifacts: vec![],
-                },
-            )
-            .await
-            .unwrap();
-
-        let result = checkpoint_episode(&memory, episode_id, "test".to_string()).await;
-        assert!(result.is_err());
-    }
-}
+#[path = "operations_tests.rs"]
+mod tests;
 
 #[cfg(test)]
-mod abstention_tests {
-    use super::*;
-    use crate::episode::Episode;
-    use crate::types::{TaskContext, TaskOutcome, TaskType};
-
-    fn create_test_episode_with_steps(step_count: usize) -> Episode {
-        let mut episode = Episode::new(
-            "Test task".to_string(),
-            TaskContext::default(),
-            TaskType::Testing,
-        );
-        for i in 1..=step_count {
-            episode.add_step(crate::episode::ExecutionStep::new(
-                i,
-                "tool".to_string(),
-                "action".to_string(),
-            ));
-        }
-        episode
-    }
-
-    #[test]
-    fn test_abstention_auto_checkpoint_created() {
-        let mut episode = create_test_episode_with_steps(5);
-        episode.outcome = Some(TaskOutcome::Abstained {
-            reason: "Tool not found".to_string(),
-            stopped_at_step: 3,
-            infeasibility_signals: vec!["404_tool".to_string()],
-        });
-        maybe_create_abstention_checkpoint(&mut episode);
-        assert_eq!(episode.checkpoints.len(), 1);
-        assert!(episode.checkpoints[0].is_abstention_checkpoint);
-        assert_eq!(episode.checkpoints[0].step_number, 3);
-        assert!(episode.checkpoints[0].label.starts_with("[ABSTAIN]"));
-    }
-
-    #[test]
-    fn test_no_checkpoint_for_zero_step_abstention() {
-        let mut episode = create_test_episode_with_steps(0);
-        episode.outcome = Some(TaskOutcome::Abstained {
-            reason: "Immediate rejection".to_string(),
-            stopped_at_step: 0,
-            infeasibility_signals: vec![],
-        });
-        maybe_create_abstention_checkpoint(&mut episode);
-        // No steps = no useful state to checkpoint
-        assert_eq!(episode.checkpoints.len(), 0);
-    }
-
-    #[test]
-    fn test_non_abstention_outcome_no_auto_checkpoint() {
-        let mut episode = create_test_episode_with_steps(5);
-        episode.outcome = Some(TaskOutcome::Failure {
-            reason: "Error".to_string(),
-            error_details: None,
-        });
-        maybe_create_abstention_checkpoint(&mut episode);
-        assert_eq!(episode.checkpoints.len(), 0);
-    }
-
-    #[test]
-    fn test_abstention_checkpoint_backward_compat_deserialization() {
-        // Old JSON without is_abstention_checkpoint and episode_id fields should deserialize with defaults
-        // and aliases should handle timestamp/label
-        let json = r#"{
-            "checkpoint_id": "00000000-0000-0000-0000-000000000000",
-            "step_number": 3,
-            "created_at": "2026-01-01T00:00:00Z",
-            "reason": "old checkpoint",
-            "salient_features_snapshot": null,
-            "note": "old note"
-        }"#;
-        let meta: CheckpointMeta = serde_json::from_str(json).unwrap();
-        assert!(!meta.is_abstention_checkpoint);
-        assert_eq!(meta.label, "old checkpoint");
-        assert_eq!(meta.step_number, 3);
-    }
-}
+#[path = "abstention_tests.rs"]
+mod abstention_tests;
