+++ memory-core/src/retrieval/cache/mod.rs
@@ -123,4 +123,4 @@ pub mod types;
 
 // Re-export public API
 pub use lru::QueryCache;
-pub use types::{CacheKey, CacheMetrics, DEFAULT_CACHE_TTL, DEFAULT_MAX_ENTRIES};
+pub use types::{CacheKey, CacheMetrics, DEFAULT_CACHE_TTL, DEFAULT_MAX_ENTRIES, normalize_tags};
